create history tab
