+++ lib/data/data.exports.dart
@@ -4,7 +4,7 @@
 // flutter pub run bin/gen_export
 library data;
 
-export 'entities/entities.exports.dart';
-export 'errors/errors.exports.dart';
-export 'exceptions/exceptions.exports.dart';
 export 'models/models.exports.dart';
+export 'exceptions/exceptions.exports.dart';
+export 'errors/errors.exports.dart';
+export 'entities/entities.exports.dart';
