+++ lib/data/models/models.exports.dart
@@ -4,12 +4,12 @@
 // flutter pub run bin/gen_export
 library models;
 
-export 'auth/auth.exports.dart';
-export 'global/global.exports.dart';
-export 'history/history.exports.dart';
-export 'money_manage/money_manage.exports.dart';
-export 'money_plan/money_plan.exports.dart';
 export 'post/post.exports.dart';
-export 'school/school.exports.dart';
+export 'money_manage/money_manage.exports.dart';
+export 'auth/auth.exports.dart';
 export 'user/user.exports.dart';
+export 'history/history.exports.dart';
 export 'wallet/wallet.exports.dart';
+export 'school/school.exports.dart';
+export 'global/global.exports.dart';
+export 'money_plan/money_plan.exports.dart';
