+++ lib/data/models/money_manage/money_manage.exports.dart
@@ -6,5 +6,5 @@ library money_manage;
 
 export 'list_money_manage_item_model.dart';
 export 'list_money_manage_model.dart';
-export 'money_manage_item_model.dart';
 export 'money_manage_model.dart';
+export 'money_manage_item_model.dart';
