+++ lib/gen_export.dart
@@ -4,13 +4,13 @@
 // flutter pub run bin/gen_export
 library gen_export;
 
+export 'ui/ui.exports.dart';
 export 'core/core.exports.dart';
-export 'data/data.exports.dart';
 export 'flavor.dart';
 export 'logic/logic.exports.dart';
-export 'main.dart';
-export 'repositories/repositories.exports.dart';
-export 'routes/routes.exports.dart';
-export 'service/service.exports.dart';
-export 'ui/ui.exports.dart';
 export 'utils/utils.exports.dart';
+export 'repositories/repositories.exports.dart';
+export 'main.dart';
+export 'service/service.exports.dart';
+export 'data/data.exports.dart';
+export 'routes/routes.exports.dart';
