+++ lib/logic/blocs/auth/auth.exports.dart
@@ -4,6 +4,6 @@
 // flutter pub run bin/gen_export
 library auth;
 
-export 'auth_bloc.dart';
 export 'auth_event.dart';
+export 'auth_bloc.dart';
 export 'auth_state.dart';
