+++ lib/logic/blocs/home/home.exports.dart
@@ -4,6 +4,6 @@
 // flutter pub run bin/gen_export
 library home;
 
+export 'home_state.dart';
 export 'home_bloc.dart';
 export 'home_event.dart';
-export 'home_state.dart';
