+++ lib/repositories/repositories.exports.dart
@@ -4,11 +4,11 @@
 // flutter pub run bin/gen_export
 library repositories;
 
-export 'history_repository.dart';
-export 'mapper/mapper.exports.dart';
-export 'money_manage_repository.dart';
 export 'money_plan_repository.dart';
-export 'school_repository.dart';
-export 'transaction_repository.dart';
+export 'mapper/mapper.exports.dart';
+export 'history_repository.dart';
 export 'user_repository.dart';
+export 'school_repository.dart';
 export 'wallet_repository.dart';
+export 'transaction_repository.dart';
+export 'money_manage_repository.dart';
