+++ lib/routes/routes.exports.dart
@@ -5,5 +5,5 @@
 library routes;
 
 export 'argument_bundle.dart';
-export 'page_path.dart';
 export 'page_router.dart';
+export 'page_path.dart';
