+++ lib/service/micro/micro.exports.dart
@@ -5,9 +5,9 @@
 library micro;
 
 export 'history_service.dart';
-export 'money_manage.dart';
 export 'money_plan_service.dart';
-export 'school_service.dart';
 export 'transaction_service.dart';
-export 'user_service.dart';
 export 'wallet_service.dart';
+export 'school_service.dart';
+export 'money_manage.dart';
+export 'user_service.dart';
