+++ lib/service/service.exports.dart
@@ -4,7 +4,7 @@
 // flutter pub run bin/gen_export
 library service;
 
-export 'locals/locals.exports.dart';
 export 'micro/micro.exports.dart';
-export 'services.dart';
+export 'locals/locals.exports.dart';
 export 'service_base.dart';
+export 'services.dart';
