+++ lib/ui/widgets/widgets.exports.dart
@@ -4,17 +4,18 @@
 // flutter pub run bin/gen_export
 library widgets;
 
-export 'background.dart';
 export 'custom_app_bar.dart';
-export 'custom_button.dart';
-export 'custom_dialog.dart';
-export 'custom_network_image.dart';
-export 'custom_numeric_keypad.dart';
-export 'custom_pin_code.dart';
+export 'merchant_item.dart';
 export 'custom_slider.dart';
+export 'history_tab.dart';
 export 'custom_tab_bar.dart';
+export 'custom_network_image.dart';
+export 'transaction_item.dart';
 export 'feature_item.dart';
 export 'loading_overlay.dart';
-export 'merchant_item.dart';
 export 'pin_code.dart';
-export 'transaction_item.dart';
+export 'custom_numeric_keypad.dart';
+export 'custom_pin_code.dart';
+export 'custom_button.dart';
+export 'custom_dialog.dart';
+export 'background.dart';
